Add transaction list in address details modal
