+++ locales/vi-VN/translation.json
@@ -318,5 +318,6 @@
   "Coming soon!": "Coming soon!",
   "Address tokens": "Address tokens",
   "Address NFTs": "Address NFTs",
-  "Your tokens will appear here.": "Your tokens will appear here."
+  "Your tokens will appear here.": "Your tokens will appear here.",
+  "Address transactions": "Address transactions"
 }
