+++ src/pages/UnlockedWallet/OverviewPage/index.tsx
@@ -17,24 +17,18 @@ along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
 import { motion } from 'framer-motion'
-import { useState } from 'react'
 import styled from 'styled-components'
 
 import { fadeIn } from '@/animations'
+import TransactionsList from '@/components/TransactionsList'
 import { useAppSelector } from '@/hooks/redux'
-import ModalPortal from '@/modals/ModalPortal'
-import TransactionDetailsModal from '@/modals/TransactionDetailsModal'
 import AddressesContactsList from '@/pages/UnlockedWallet/OverviewPage/AddressesContactsList'
 import AmountsOverviewPanel from '@/pages/UnlockedWallet/OverviewPage/AmountsOverviewPanel'
 import TokensNFTsList from '@/pages/UnlockedWallet/OverviewPage/TokensNFTsList'
-import TransactionList from '@/pages/UnlockedWallet/OverviewPage/TransactionList'
 import { UnlockedWalletPanel } from '@/pages/UnlockedWallet/UnlockedWalletLayout'
-import { AddressConfirmedTransaction } from '@/types/transactions'
 
 const OverviewPage = () => {
   const isLoadingData = useAppSelector((state) => state.addresses.loading)
-
-  const [selectedTransaction, setSelectedTransaction] = useState<AddressConfirmedTransaction>()
 
   return (
     <motion.div {...fadeIn}>
@@ -44,17 +38,7 @@ const OverviewPage = () => {
           <TokensNFTsList showTokens showNfts />
           <AddressesContactsList limit={5} />
         </Row>
-        <TransactionList onTransactionClick={setSelectedTransaction} limit={5} />
-        <ModalPortal>
-          {selectedTransaction?.hash}
-          {selectedTransaction && (
-            <TransactionDetailsModal
-              address={selectedTransaction.address}
-              transaction={selectedTransaction}
-              onClose={() => setSelectedTransaction(undefined)}
-            />
-          )}
-        </ModalPortal>
+        <TransactionsList limit={5} />
       </UnlockedWalletPanel>
     </motion.div>
   )
